desacoplamento do resource query e view to json
