+++ api-tools/src/main/java/com/eoscode/springapitools/resource/ViewToJson.java
@@ -1,0 +1,9 @@
+package com.eoscode.springapitools.resource;
+
+import com.eoscode.springapitools.data.filter.ViewDefinition;
+
+public interface ViewToJson {
+
+    <T> String toJson(ViewDefinition viewDefinition, T result);
+
+}
