+++ api-tools/src/main/java/com/eoscode/springapitools/resource/exception/FieldMessage.java
@@ -5,12 +5,10 @@ import java.io.Serializable;
 public class FieldMessage implements Serializable {
 
     private static final long serialVersionUID = 1L;
-
     private String fieldName;
     private String message;
 
-    public FieldMessage() {
-    }
+    public FieldMessage() {}
 
     public FieldMessage(String fieldName, String message) {
         super();
